requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
scikit-image==0.21.0
matplotlib==3.8.0
pandas==2.1.1
openpyxl==3.1.2
pymupdf==1.23.5
Pillow==10.0.1
numpy==1.26.0